feat: 1차 리뷰 반영
